Moved util.collections to internal
