+++ checkstyle/suppressions.xml
@@ -99,6 +99,9 @@
     <suppress checks="MethodCount" files="com/hazelcast/concurrent/lock/LockStoreImpl"/>
     <suppress checks="Todo" files="com/hazelcast/concurrent/lock/ConditionImpl"/>
 
+    <!-- Collection -->
+    <suppress checks="MethodCount" files="com/hazelcast/collection/impl/collection/CollectionContainer"/>
+
     <!-- OSGI -->
     <suppress checks="MethodCount" files="com/hazelcast/osgi/impl/HazelcastOSGiInstanceImpl"/>
 
@@ -243,12 +246,11 @@
     <suppress checks="ClassFanOutComplexity|MethodCount" files="com/hazelcast/executor/impl/ExecutorServiceProxy"/>
 
     <!-- Collections (incl. Agrona backport) -->
-    <suppress checks="JavadocType" files="com/hazelcast/util/collection/\w*2ObjectHashMap"/>
-    <suppress checks="MethodCount" files="com/hazelcast/util/collection/\w*Hash\w*"/>
-    <suppress checks="MethodCount" files="com/hazelcast/collection/impl/collection/CollectionContainer"/>
-    <suppress checks="MagicNumber|Header" files="com/hazelcast/util/collection/"/>
+    <suppress checks="JavadocType" files="com/hazelcast/internal/util/collection/\w*2ObjectHashMap"/>
+    <suppress checks="MethodCount" files="com/hazelcast/internal/util/collection/\w*Hash\w*"/>
+    <suppress checks="MagicNumber|Header" files="com/hazelcast/internal/util/collection/"/>
+    <suppress checks="SuperClone" files="com/hazelcast/internal/util/collection/InflatableSet"/>
     <suppress checks="Header" files="com/hazelcast/util/QuickMath"/>
-    <suppress checks="SuperClone" files="com/hazelcast/util/collection/InflatableSet"/>
 
     <!-- Queue -->
     <suppress checks="MethodCount" files="com/hazelcast/collection/impl/queue/QueueContainer"/>
