+++ hazelcast/src/main/java/com/hazelcast/internal/util/collection/IntIterator.java
@@ -15,47 +15,48 @@
  * limitations under the License.
  */
 
-package com.hazelcast.util.collection;
+package com.hazelcast.internal.util.collection;
 
 import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
 
 import java.util.Iterator;
+import java.util.NoSuchElementException;
 
 /**
  * An iterator for a sequence of primitive integers.
  */
-@SuppressFBWarnings("EI2")
-public class LongIterator implements Iterator<Long> {
-    private final long missingValue;
-    private final long[] values;
+public class IntIterator implements Iterator<Integer> {
+    private final int missingValue;
+    private final int[] values;
 
     private int position;
 
     /**
-     * Construct an {@link Iterator} over an array of primitives longs.
+     * Construct an {@link Iterator} over an array of primitives ints.
      *
      * @param missingValue to indicate the value is missing, i.e. not present or null.
      * @param values       to iterate over.
      */
-    public LongIterator(final long missingValue, final long[] values) {
+    @SuppressFBWarnings(value = "EI2", justification =
+            "This is flyweight over caller's array, so no copying")
+    public IntIterator(final int missingValue, final int[] values) {
         this.missingValue = missingValue;
         this.values = values;
+        this.position = -1;
     }
 
     public boolean hasNext() {
-        final long[] values = this.values;
+        final int[] values = this.values;
         while (position < values.length) {
-            if (values[position] != missingValue) {
+            if (position >= 0 && values[position] != missingValue) {
                 return true;
             }
-
             position++;
         }
-
         return false;
     }
 
-    public Long next() {
+    public Integer next() {
         return nextValue();
     }
 
@@ -64,12 +65,15 @@ public class LongIterator implements Iterator<Long> {
     }
 
     /**
-     * Strongly typed alternative of {@link Iterator#next()} not to avoid boxing.
+     * Strongly typed alternative of {@link Iterator#next()} to avoid boxing.
      *
-     * @return the next long value.
+     * @return the next int value.
      */
-    public long nextValue() {
-        final long value = values[position];
+    public int nextValue() {
+        if (!hasNext()) {
+            throw new NoSuchElementException();
+        }
+        final int value = values[position];
         position++;
         return value;
     }
