+++ hazelcast/src/main/java/com/hazelcast/query/impl/predicates/VisitorUtils.java
@@ -20,7 +20,7 @@ import com.hazelcast.query.Predicate;
 import com.hazelcast.query.VisitablePredicate;
 import com.hazelcast.query.impl.Indexes;
 
-import static com.hazelcast.util.collection.ArrayUtils.createCopy;
+import static com.hazelcast.internal.util.collection.ArrayUtils.createCopy;
 
 public final class VisitorUtils {
